[REFACTOR]: 범주별 함수 리스트 조회 API - URL 통일을 위한 컨트롤러 리팩토링
